[refactor] 기획 변경으로 인한 RepoDataSource 삭제 및 폴더 구조 수정
- api 폴더 생성
- data/remote/api에서 Service(api) 관리
