+++ data/src/main/java/com/example/data/remote/api/UserService.kt
@@ -1,4 +1,4 @@
-package com.example.data.remote
+package com.example.data.remote.api
 
 import com.example.data.remote.response.user.UserResponse
 import retrofit2.http.GET
